+++ docs/main/changelog.rst
@@ -9,8 +9,8 @@ follows `Semantic versioning`_
 
 Development version
 -------------------
-- Improve async mode exceptions handling to prevent infinite hanging when exception occurs.
-- Fix double printing of exception when resource initialization causes an error.
+- Improve async mode exceptions handling.
+- Fix double printing of exception when resource initialization causes an error in async mode.
 
 4.23.1
 ------